:label: Modification du type de prixEuro int => float

+++ src/Models/article.class.php
@@ -95,9 +95,9 @@ class Article {
     /**
      * @brief Le prix en euro de l'article
      *
-     * @var int|null
+     * @var float|null
      */
-    private ?int $priceEuro;
+    private ?float $priceEuro;
     /**
      * @brief Le chemin d'accès de l'image
      *
@@ -115,10 +115,10 @@ class Article {
      * @param ArticleType|null $type Le type de l'article
      * @param string|null $description La description de l'article
      * @param int|null $pricePoint Le prix en point de l'article
-     * @param int|null $priceEuro Le prix en euro de l'article
+     * @param float|null $priceEuro Le prix en euro de l'article
      * @param string|null $filePath Le chemin de l'image de l'article
      */
-    public function __construct(?int $id = null, ?DateTime $createdAt = null, ?DateTime $updatedAt = null, ?string $name = null, ?ArticleType $type = null, ?string $description = null , ?int $pricePoint = null, ?int $priceEuro = null, ?string $filePath = null)
+    public function __construct(?int $id = null, ?DateTime $createdAt = null, ?DateTime $updatedAt = null, ?string $name = null, ?ArticleType $type = null, ?string $description = null , ?int $pricePoint = null, ?float $priceEuro = null, ?string $filePath = null)
     {
         $this->id = $id;
         $this->createdAt = $createdAt;
@@ -282,9 +282,9 @@ class Article {
     /**
      * @brief Retourne le prix en euro de l'article
      *
-     * @return int|null Le prix en euro de l'article qui est retourné
+     * @return float|null Le prix en euro de l'article qui est retourné
      */
-    public function getPriceEuro(): ?int
+    public function getPriceEuro(): ?float
     {
         return $this->priceEuro;
     }
@@ -292,10 +292,10 @@ class Article {
     /**
      * @brief Modifie le prix en euro de l'article
      *
-     * @param int|null $priceEuro Le nouveau prix de l'article
+     * @param float|null $priceEuro Le nouveau prix de l'article
      * @return void
      */
-    public function setPriceEuro(?int $priceEuro): void
+    public function setPriceEuro(?float $priceEuro): void
     {
         $this->priceEuro = $priceEuro;
     }
